requirements.txt
requests==2.31.0
urllib3==1.26
pytz
google-cloud-bigquery